Fix to make redirects to other page work in IE (11)
Replaced window.location.assign() with $location.path()
window.location.assign() caused the following error in IE:
"Error: [$rootScope:infdig] 10 digest() iterations reached. Aborting!"
and the redirect took long time.

Note: a possible drawback with the current solotion is that the page seems to be completely reloaded now which it was not before (as I understand it).